Refactor listings-importer library (#378)
* Move createUnitsArray

* Pass error up

* Formatting change
