+++ backend/core/scripts/listings-importer.ts
@@ -2,9 +2,21 @@ import * as client from "../types/src/backend-swagger"
 import axios from "axios"
 import { ListingCreate, serviceOptions } from "../types/src/backend-swagger"
 import { ListingStatus } from "../src/listings/types/listing-status-enum"
+import { UnitStatus } from "../src/units/types/unit-status-enum"
 
 // NOTE: This script relies on any logged-in users having permission to create
 // listings and properties (defined in backend/core/src/auth/authz_policy.csv)
+
+export function createUnitsArray(type: string, number: number) {
+  const units = []
+  for (let unit_index = 0; unit_index < number; unit_index++) {
+    units.push({
+      unitType: type,
+      status: UnitStatus.unknown,
+    })
+  }
+  return units
+}
 
 function uploadPreferences(listing) {
   const preferencesService = new client.PreferencesService()
@@ -28,9 +40,7 @@ async function uploadListing(listing: ListingCreate) {
       body: listing,
     })
   } catch (e) {
-    console.log(listing)
-    console.log(e.response.data.message)
-    process.exit(1)
+    throw new Error(e.response.data.message)
   }
 }
